Split up logger and widget
This allows us to use the logger from the wtf directory
For example when trying to debug sort ordering in focus_tracker
