+++ modules/logger/settings.go
